Fix repeated fires. Primary info capture. Use grocerry raw data sent with webRequests
Only one request is made for this info by the grocery app. So, we fix
the duplicate event fires (Yay!).

We fetch the response that the request had already made.
This is a hack to wordaround WebRequests not giving response data.

We must be careful and reject any requests made from the extension.
If we don't we'll request forever (Don't want to be blocked by the website for abuse).